array dispatch helper
